fixing check fail1

+++ 0-setup_web_static.sh
@@ -21,9 +21,9 @@ echo "
    </body>
 </html>" > /data/web_static/releases/test/index.html
 
-# if [ -L /data/web_static/current ]; then
-# 	sudo rm /data/web_static/current
-# fi
+if [ -L /data/web_static/current ]; then
+	sudo rm /data/web_static/current
+fi
 sudo ln -s /data/web_static/releases/test/  /data/web_static/current
 
 #recursivery Give ownership of the /data/ folder to the ubuntu user AND group
@@ -32,7 +32,9 @@ sudo chown -R ubuntu: /data/
 #add a location block for /hbnb_static in sites-available/default file
 sudo sed -i '/server_name _;/a\\tlocation \/hbnb_static {\n\t\talias \/data\/web_static\/current;\n\t}' /etc/nginx/sites-available/default
 
-sudo rm /etc/nginx/sites-enabled/default
+if [ -L /etc/nginx/sites-enabled/default ]; then
+	sudo rm /etc/nginx/sites-enabled/default
+fi
 
 sudo ln -s /etc/nginx/sites-available/default  /etc/nginx/sites-enabled/
